--- lib/Screens/Form.dart
@@ -1,11 +1,0 @@
-import 'package:flutter/src/widgets/framework.dart';
-import 'package:flutter/src/widgets/placeholder.dart';
-
-class Formfillin extends StatelessWidget {
-  const Formfillin({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return const Placeholder();
-  }
-}
