+++ lib/main.dart
@@ -1,9 +1,7 @@
-import 'package:edd/Screens/Register.dart';
 import 'package:edd/Screens/Splashscreen.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:easy_localization/easy_localization.dart';
-
 
 // void main() {
 //   runApp(const MyApp());
@@ -15,17 +13,16 @@ Future main() async {
   WidgetsFlutterBinding.ensureInitialized();
   await EasyLocalization.ensureInitialized();
   runApp(EasyLocalization(
-    supportedLocales: [
+    supportedLocales: const [
       Locale('en', 'US'),
       Locale('am', 'ETH'),
     ],
     path: 'assets/translations',
     saveLocale: true,
-    fallbackLocale: Locale('en', 'US'),
-    child: MyApp(),
+    fallbackLocale: const Locale('en', 'US'),
+    child: const MyApp(),
   ));
 }
-
 
 final navigatorKey = GlobalKey<NavigatorState>();
 
@@ -38,13 +35,12 @@ class MyApp extends StatelessWidget {
     return MaterialApp(
       navigatorKey: navigatorKey,
       title: 'EDD',
-      home: Splashscreen(),
+      home: const Splashscreen(),
       debugShowCheckedModeBanner: false,
       // loclization
       localizationsDelegates: context.localizationDelegates,
       supportedLocales: context.supportedLocales,
       locale: context.locale,
-      
-    ); 
+    );
   }
 }
